pyproject.toml
[project]
name = "datahub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-jazzmin>=3.0.1",
    "django-simple-history>=3.11.0",
    "django-stubs>=5.2.8",
]
